+++ tests/MutationFilters/MutationFilterTests.cpp
@@ -4,6 +4,7 @@
 #include "mull/BitcodeLoader.h"
 #include "mull/Config/Configuration.h"
 #include "mull/Filter.h"
+#include "mull/MutationFilters/FilePathFilter.h"
 #include "mull/MutationFilters/NoDebugInfoFilter.h"
 #include "mull/MutationsFinder.h"
 #include "mull/Mutators/MathAddMutator.h"
@@ -83,6 +84,165 @@ TEST(NoDebugInfoFilter, withouDebugInfo) {
   NoDebugInfoFilter mutationFilter;
   for (auto point : points) {
     if (!mutationFilter.shouldSkip(point)) {
+      filteredPoints.push_back(point);
+    }
+  }
+
+  ASSERT_EQ(filteredPoints.size(), size_t(0));
+}
+
+TEST(FilePathFilter, doesNotFilterEmpty) {
+  llvm::LLVMContext context;
+  BitcodeLoader loader;
+  auto bitcodeFile = loader.loadBitcodeAtPath(
+      fixtures::mutation_filters_file_path_some_test_file_name_bc_path(),
+      context);
+  auto module = bitcodeFile->getModule();
+
+  std::vector<std::unique_ptr<Bitcode>> bitcode;
+  bitcode.push_back(std::move(bitcodeFile));
+  Program program({}, {}, std::move(bitcode));
+  Configuration configuration;
+
+  std::vector<std::unique_ptr<Mutator>> mutators;
+  mutators.emplace_back(llvm::make_unique<MathAddMutator>());
+  MutationsFinder finder(std::move(mutators), configuration);
+  Filter filter;
+
+  std::vector<std::unique_ptr<Testee>> testees;
+  for (auto &function : *module) {
+    testees.emplace_back(llvm::make_unique<Testee>(&function, nullptr, 1));
+  }
+  auto mergedTestees = mergeTestees(testees);
+  std::vector<MutationPoint *> points =
+      finder.getMutationPoints(program, mergedTestees, filter);
+
+  ASSERT_NE(points.size(), size_t(0));
+
+  std::vector<MutationPoint *> filteredPoints;
+  FilePathFilter filePathFilter;
+  for (auto point : points) {
+    if (!filePathFilter.shouldSkip(point)) {
+      filteredPoints.push_back(point);
+    }
+  }
+
+  ASSERT_EQ(filteredPoints.size(), points.size());
+}
+
+TEST(FilePathFilter, doesNotFilterMismatch) {
+  llvm::LLVMContext context;
+  BitcodeLoader loader;
+  auto bitcodeFile = loader.loadBitcodeAtPath(
+      fixtures::mutation_filters_file_path_some_test_file_name_bc_path(),
+      context);
+  auto module = bitcodeFile->getModule();
+
+  std::vector<std::unique_ptr<Bitcode>> bitcode;
+  bitcode.push_back(std::move(bitcodeFile));
+  Program program({}, {}, std::move(bitcode));
+  Configuration configuration;
+
+  std::vector<std::unique_ptr<Mutator>> mutators;
+  mutators.emplace_back(llvm::make_unique<MathAddMutator>());
+  MutationsFinder finder(std::move(mutators), configuration);
+  Filter filter;
+
+  std::vector<std::unique_ptr<Testee>> testees;
+  for (auto &function : *module) {
+    testees.emplace_back(llvm::make_unique<Testee>(&function, nullptr, 1));
+  }
+  auto mergedTestees = mergeTestees(testees);
+  std::vector<MutationPoint *> points =
+      finder.getMutationPoints(program, mergedTestees, filter);
+
+  ASSERT_NE(points.size(), size_t(0));
+
+  std::vector<MutationPoint *> filteredPoints;
+  FilePathFilter filePathFilter;
+  filePathFilter.exclude("irrelevant");
+  for (auto point : points) {
+    if (!filePathFilter.shouldSkip(point)) {
+      filteredPoints.push_back(point);
+    }
+  }
+
+  ASSERT_EQ(filteredPoints.size(), points.size());
+}
+
+TEST(FilePathFilter, filtersPlainString) {
+  llvm::LLVMContext context;
+  BitcodeLoader loader;
+  auto bitcodeFile = loader.loadBitcodeAtPath(
+      fixtures::mutation_filters_file_path_some_test_file_name_bc_path(),
+      context);
+  auto module = bitcodeFile->getModule();
+
+  std::vector<std::unique_ptr<Bitcode>> bitcode;
+  bitcode.push_back(std::move(bitcodeFile));
+  Program program({}, {}, std::move(bitcode));
+  Configuration configuration;
+
+  std::vector<std::unique_ptr<Mutator>> mutators;
+  mutators.emplace_back(llvm::make_unique<MathAddMutator>());
+  MutationsFinder finder(std::move(mutators), configuration);
+  Filter filter;
+
+  std::vector<std::unique_ptr<Testee>> testees;
+  for (auto &function : *module) {
+    testees.emplace_back(llvm::make_unique<Testee>(&function, nullptr, 1));
+  }
+  auto mergedTestees = mergeTestees(testees);
+  std::vector<MutationPoint *> points =
+      finder.getMutationPoints(program, mergedTestees, filter);
+
+  ASSERT_NE(points.size(), size_t(0));
+
+  std::vector<MutationPoint *> filteredPoints;
+  FilePathFilter filePathFilter;
+  filePathFilter.exclude("some_test");
+  for (auto point : points) {
+    if (!filePathFilter.shouldSkip(point)) {
+      filteredPoints.push_back(point);
+    }
+  }
+
+  ASSERT_EQ(filteredPoints.size(), size_t(0));
+}
+
+TEST(FilePathFilter, filtersWithRegex) {
+  llvm::LLVMContext context;
+  BitcodeLoader loader;
+  auto bitcodeFile = loader.loadBitcodeAtPath(
+      fixtures::mutation_filters_file_path_some_test_file_name_bc_path(),
+      context);
+  auto module = bitcodeFile->getModule();
+
+  std::vector<std::unique_ptr<Bitcode>> bitcode;
+  bitcode.push_back(std::move(bitcodeFile));
+  Program program({}, {}, std::move(bitcode));
+  Configuration configuration;
+
+  std::vector<std::unique_ptr<Mutator>> mutators;
+  mutators.emplace_back(llvm::make_unique<MathAddMutator>());
+  MutationsFinder finder(std::move(mutators), configuration);
+  Filter filter;
+
+  std::vector<std::unique_ptr<Testee>> testees;
+  for (auto &function : *module) {
+    testees.emplace_back(llvm::make_unique<Testee>(&function, nullptr, 1));
+  }
+  auto mergedTestees = mergeTestees(testees);
+  std::vector<MutationPoint *> points =
+      finder.getMutationPoints(program, mergedTestees, filter);
+
+  ASSERT_NE(points.size(), size_t(0));
+
+  std::vector<MutationPoint *> filteredPoints;
+  FilePathFilter filePathFilter;
+  filePathFilter.exclude("some.*name");
+  for (auto point : points) {
+    if (!filePathFilter.shouldSkip(point)) {
       filteredPoints.push_back(point);
     }
   }