+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Print more info when original test fails #549
  - Reporting JSON to Mutation Testing Elements (initial support) #517, #507
+ - Introduce `-exclude-path` option to filter mutations #553
 
 ## [0.4.0] - 11 Aug 2019